Refactor for dalle name fix
